+++ client/src/hooks/useJwt.js
@@ -1,6 +1,13 @@
 import jwtDecode from 'jwt-decode';
 // import React from 'react';
 // not technically a "hook" -- not stateful and no side effects but yeah
+const decode = jwt => {
+    try {
+        return jwtDecode(jwt);
+    } catch (e) {
+        return '';
+    }
+};
 export default function useJwt() {
     const jwt = localStorage.getItem('jwt');
     // React.useEffect(() => {
@@ -10,6 +17,6 @@ export default function useJwt() {
     //     };
     // });
     // don't decode unless the token exists
-    const rawToken = jwt ? jwtDecode(jwt) : null;
+    const rawToken = jwt ? decode(jwt) : null;
     return [jwt, rawToken];
 }
