+++ client/src/pages/index.jsx
@@ -3,9 +3,10 @@ import { Route, Redirect, Switch } from 'react-router-dom';
 import Sessions from './Sessions';
 import Chat from './Chat';
 import Logout from './Logout';
-import Login from './LoginPage';
+import Login from './Login';
 import Layout from '../layout';
 import LoggedIn from '../components/LoggedIn';
+import LoginTemp from './LoginTemp';
 
 export default function Routes() {
     return (
@@ -24,6 +25,9 @@ export default function Routes() {
             </Route>
             <Route exact path='/login'>
                 <Login />
+            </Route>
+            <Route exact path='/login-temporary'>
+                <LoginTemp />
             </Route>
             <Route exact path='/logout'>
                 <Logout />
